adding docs, refactoring and tests
